refactor: reorganize DTO classes and remove unused files
