--- auth/src/main/java/com/ndhuy/auth/authentication/application/dto/AddSessionJwtOut.java
@@ -1,16 +1,0 @@
-package com.ndhuy.auth.authentication.application.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class AddSessionJwtOut {
-    private String jwtSession;
-}
