--- auth/src/main/java/com/ndhuy/auth/authentication/application/dto/AddSessionUserIn.java
@@ -1,23 +1,0 @@
-package com.ndhuy.auth.authentication.application.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@ToString
-public class AddSessionUserIn {
-
-    @NotNull
-    private String username;
-    @NotNull
-    private String jwtSession;
-}
