--- auth/src/main/java/com/ndhuy/auth/authentication/application/dto/GetSesssionAuthOut.java
@@ -1,18 +1,0 @@
-package com.ndhuy.auth.authentication.application.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class GetSesssionAuthOut {
-    private String jwtSession;
-    private String issueAt;
-    private String expiresAt;
-}
