--- auth/src/main/java/com/ndhuy/auth/authentication/application/dto/PermissionOut.java
@@ -1,19 +1,0 @@
-package com.ndhuy.auth.authentication.application.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PermissionOut {
-   private String jwt;
-   private String jwtRefresh;
-   private String issueAt;
-   private String expiresAt;
-}
