+++ auth/src/main/java/com/ndhuy/auth/authentication/application/dto/RefrestSessionDto.java
@@ -1,0 +1,13 @@
+package com.ndhuy.auth.authentication.application.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class RefrestSessionDto {
+    @Getter
+    @Setter
+    public static class RefrestSessionOut extends PermissionDto.PermissionIn {
+
+    }
+
+}
