--- auth/src/main/java/com/ndhuy/auth/authentication/application/dto/RefrestSessionOut.java
@@ -1,10 +1,0 @@
-package com.ndhuy.auth.authentication.application.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RefrestSessionOut extends PermissionOut{
-    
-}
