+++ auth/src/main/java/com/ndhuy/auth/authentication/application/dto/RemoveSessionDto.java
@@ -1,0 +1,12 @@
+package com.ndhuy.auth.authentication.application.dto;
+
+import lombok.Builder;
+
+public class RemoveSessionDto {
+
+    @Builder
+    public static class RemoveSessionJwtOut {
+        boolean isDone;
+    }
+
+}
