--- auth/src/main/java/com/ndhuy/auth/authentication/application/dto/RemoveSessionJwtOut.java
@@ -1,8 +1,0 @@
-package com.ndhuy.auth.authentication.application.dto;
-
-import lombok.Builder;
-
-@Builder
-public class RemoveSessionJwtOut {
-    boolean isDone;
-}
