+++ auth/src/main/java/com/ndhuy/auth/authentication/application/rest/AuthenticationRestPublic.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ndhuy.auth.authentication.application.dto.PermissionIn;
-import com.ndhuy.auth.authentication.application.dto.PermissionOut;
+import com.ndhuy.auth.authentication.application.dto.PermissionDto.PermissionIn;
+import com.ndhuy.auth.authentication.application.dto.PermissionDto.PermissionOut;
 import com.ndhuy.auth.authentication.application.service.PermissionService;
 import com.ndhuy.auth.user.application.dto.OpenAccountIn;
 import com.ndhuy.auth.user.application.dto.OpenAccountOut;
