+++ auth/src/main/java/com/ndhuy/auth/authentication/application/service/RefreshSessionService.java
@@ -1,6 +1,6 @@
 package com.ndhuy.auth.authentication.application.service;
 
-import com.ndhuy.auth.authentication.application.dto.RefrestSessionOut;
+import com.ndhuy.auth.authentication.application.dto.RefrestSessionDto.RefrestSessionOut;
 
 public interface RefreshSessionService {
     RefrestSessionOut doMain(String jwt);
